Moves the fortls interface
